+++ rootdriver/conversation.py
@@ -2,8 +2,9 @@ from .types import Message, Role
 from .utils import get_iso_timestamp, build_system_message
 
 
+
 class Conversation:
-    def __init__(self, system_prompt: str = None):
+    def __init__(self,  system_prompt: str = None):
         self.messages: list[Message] = []
         self.system_prompt = system_prompt
         if self.system_prompt:
@@ -16,28 +17,7 @@ class Conversation:
     def append_many(self, messages: list[Message]) -> "Conversation":
         self.messages.extend(messages)
         return self
-
-    # def append_user(self, content: str) -> "Conversation":
-    #     self.messages.append(Message(role="user", content=content, created_at=get_iso_timestamp()))
-    #     return self
-    #
-    # def append_assistant(self, content: str | None = None, tool_calls: list | None = None) -> "Conversation":
-    #     msg = Message(role="assistant", content=content, created_at=get_iso_timestamp())
-    #     if tool_calls:
-    #         msg.tool_calls = tool_calls
-    #     self.messages.append(msg)
-    #     return self
-    #
-    # def append_tool(self, tool_call_id: str, content: str) -> "Conversation":
-    #     self.messages.append(Message(
-    #         role="tool",
-    #         tool_call_id=tool_call_id,
-    #         content=content,
-    #         created_at=get_iso_timestamp()
-    #     ))
-    #     return self
-
-    def append_system(self, content: str) -> "Conversation":
+    def append_system_message(self, content: str) -> "Conversation":
         self.messages.append(Message(role=Role.SYSTEM, content=content, created_at=get_iso_timestamp()))
         return self
     def update_message(self, messages: list[Message]) -> "Conversation":
@@ -48,10 +28,10 @@ class Conversation:
         self.messages.pop(index)
         return self
 
+
     def get_messages(self) -> list[Message]:
         '''返回 list[message<obj>]'''
         return self.messages
-
     def get_messages_in_list(self) -> list[dict]:
         '''返回 字典加列表组成的message（可转成json用于网络、跨语言传输） 组成的列表'''
         return [m.model_dump(exclude_none=False) for m in self.messages]
@@ -60,6 +40,8 @@ class Conversation:
     def get_message_in_list_by_index(self, index:int=-1) -> list[dict]:
         """据索引返回当前消息列表的"""
         return self.messages[index].model_dump(exclude_none=False)
+
+
     @classmethod
     def from_dict_list(cls, messages: list[dict]) -> "Conversation":
         conv = cls()