+++ rootdriver/conversation_repo.py
@@ -1,43 +1,81 @@
+import uuid
+
 from .db import JsonDB
-from .conversation import Conversation
-from .exceptions import CheckpointNotFoundError, DBAgentIDNotFoundError
+from .types import Message
+from .constants import  DEFAULT_AGENT_DB_PATH, DEFAULT_AGENT_STATE_AUTO_SAVE_NAME
+class DBOpt:
+    """数据库持久化层"""
+
+    def __init__(self, db: JsonDB, repo_id: str):
+        self.db = db
+        self.repo_id = repo_id
+
+    def update(self, messages:list[Message], checkpoint_name: str) -> 'DBOpt':
+        """全量覆写 conversation 到数据库。"""
+        self.db.update(
+            checkpoint_name,
+            self.repo_id,
+            [message.model_dump() for message in messages],
+        )
+        return self
+
+    def append(self, message: Message, checkpoint_name: str) -> "DBOpt":
+        """追加单条 message 到数据库。"""
+        self.db.append(checkpoint_name, self.repo_id, message.model_dump())
+        return self
+
+    def sync_append(self, messages:list[Message], checkpoint_name: str, start_index: int = 0) -> "DBOpt":
+        """增量追加从 start_index 起的新消息到数据库。"""
+        for i in range(start_index, len(messages)):
+            self.db.append(checkpoint_name, self.repo_id, messages[i].model_dump())
+        return self
+
+    def get(self, checkpoint_name: str) -> list[Message]:
+        """从数据库返回 messages 列表。"""
+        return [Message.model_load(msg_dict) for msg_dict in self.db.read(checkpoint_name, self.repo_id)]
+
+
+class BufferOpt:
+    """内存快照层"""
+    def __init__(self):
+        self.checkpoints = {}
+
+    def update(self,messages:list[Message],  checkpoint_name: str) -> "BufferOpt":
+        """保存 conversation 拷贝到内存快照。"""
+        self.checkpoints[checkpoint_name] = messages
+        return self
+
+    def get(self, checkpoint_name: str) -> list[Message]|None:
+        """从内存快照加载 messages。"""
+        messages = self.checkpoints.get(checkpoint_name)
+        return messages
+
+    def remove(self, checkpoint_name: str) -> "BufferOpt":
+        """删除内存快照。"""
+        if checkpoint_name in self.checkpoints:
+            del self.checkpoints[checkpoint_name]
+        return self
 
 
 class ConversationRepo:
-    """Conversation 的持久化仓库。"""
+    """Conversation 的持久化仓库。绑定 conversation/db/repo_id，一对一管理一个 Conversation 的持久化。"""
 
-    def __init__(self, db: JsonDB):
-        self.db = db
+    def __init__(self, db_opt: DBOpt, buffer_opt: BufferOpt, repo_id: str):
+        self.db_opt = db_opt
+        self.buffer_opt = buffer_opt
+        self.repo_id = repo_id
 
-    def save(self, conversation: Conversation, name: str, agent_id: str, start_index:int=0, end_index:int=None) -> None:
-        """全量保存 conversation当前数据 到数据库。
-        :param end_index: 不填默认最后一个索引
+    @classmethod
+    def create(cls, db_path:str = DEFAULT_AGENT_DB_PATH, repo_id:str =  None) -> "ConversationRepo":
         """
-        if end_index is None:
-            end_index = len(conversation.messages) - 1
-        self.db.update(name, agent_id, conversation.get_messages_in_list()[start_index:end_index])
-    def append(self, conversation: Conversation, name: str, agent_id: str, index) -> None:
-        message = conversation.get_messages_in_list()[index]
-        self.db.append(name, agent_id, message)
-    def auto_append(self, conversation: Conversation, name: str, agent_id: str, start_index: int=0) -> None:
-        """增量追加，从 start 索引起的 message 追加到数据库。"""
-        messages = conversation.get_messages_in_list()
-        for i in range(start_index, len(messages)):
-            self.db.append(name, agent_id, messages[i])
+        :param db_path:
+        :param repo_id: 默认会自定生成 uuid4
+        :param sync_save: 是否自定保存
+        :return:
+        """
+        repo_id = repo_id or str(uuid.uuid4())
+        db = JsonDB(db_path)
+        db_opt = DBOpt(db, repo_id)
+        buffer_opt = BufferOpt()
+        return cls(db_opt, buffer_opt, repo_id)
 
-    def load(self, conversation:Conversation, name: str, agent_id: str) -> Conversation:
-        """从数据库加载 conversation，不存在则抛异常。"""
-        load_conv = self.get(name, agent_id)
-        if load_conv is None:
-            load_conv = Conversation.from_messages_list_to_conversation([])
-        return conversation.update_message(load_conv.get_messages())
-
-    def get(self, name: str, agent_id: str) -> Conversation | None:
-        """从数据库加载 conversation，不存在返回 None。"""
-        try:
-            messages = self.db.read(name, agent_id)
-        except CheckpointNotFoundError:
-            return None
-        except DBAgentIDNotFoundError:
-            return None
-        return Conversation.from_dict_list(messages)